requirements.txt
streamlit
pandas
supabase
langgraph
google-genai
psycopg2-binary
python-dotenv
pdfplumber
reportlab